yolo-live/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.3.0
convex==0.6.0
cycler==0.12.1
Cython==3.0.11
deep-sort-realtime==1.3.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.0
flake8==7.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.0
inflect==7.3.1
isort==4.3.21
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
jiter==0.6.1
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.3.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
openai==1.51.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
tb-nightly==2.18.0a20240909
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tomli==2.0.1
torch==2.4.1
-e git+https://github.com/KaiyangZhou/deep-person-reid.git@566a56a2cb255f59ba75aa817032621784df546a#egg=torchreid
torchvision==0.19.1
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.17
ultralytics-thop==2.0.6
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.20.1
